% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVToGPX_track.R
\name{CSVToGPX_track}
\alias{CSVToGPX_track}
\title{CSVToGPX_track}
\usage{
CSVToGPX_track(inCSV, outGPX)
}
\arguments{
\item{inCSV}{filename of csv containing track data to be processed. Must
include the following columns:
uid, startLat, startLon, stopLat, stopLon
where uid is a unique track identifier, startLat/startLon is the track
start location and stopLat/stopLon is the track end location.
All locations should be in decimal degrees.
Can optionally include a startDateTime and stopDateTime column. Assumes
UTC timezone for GPX output.}

\item{outGPX}{fullpath filename of gpx to be written
example: outGPX = './crputils/exampleData/exampleVesselTrack.gpx'}
}
\value{
none, will write a file
}
\description{
convert a .csv of vessel track segments to a .gpx file.

The input data.frame must have 1 column 'uid' containing a unique
identifier for that track/segment, 2 columns for start latitude and
longitude, and 2 columns for stop latitude and longitude.
Additional columns for cruise, date, vessel name, etc will be ignored

This is a modified version of \code{trackToGPX.R} that is included in the
'cruise-maps-live' repository that tries to generalize the function to
convert any input data.frame, not just the output from a .das file.
}
\author{
Selene Fregosi
}
