% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASTrackToGPX.R
\name{DASTrackToGPX}
\alias{DASTrackToGPX}
\title{DASTrackToGPX}
\usage{
DASTrackToGPX(et, outGPX)
}
\arguments{
\item{et}{data.frame of effort as tracks, can be 'et' cumulative over a
survey or 'et' for just a single DAS}

\item{outGPX}{fullpath filename to save
example: outGPX = paste0('newEffortTracks_', dasName, '_', Sys.Date(), '.gpx}
}
\value{
none, will write a file
}
\description{
create a .gpx file based on the effort track data recorded in
.DAS files. The track data needs to have been extracted from the
.DAS file using \code{extractTrack()} which creates the 'et' dataframe

Track segments within days are plotted separately but are connected. Tracks
across days are not connected.

This was generalized from the cruise-maps-live repository's trackToGPX()

author: Selene Fregosi selene.fregosi \link{at} noaa.gov
date: 17 April 2024
}
\examples{
load('./cruise-maps-live/data/2023_leg01_OES/compiledEffortTracks_2023_leg01_OES.Rda')
trackToGPX(et, './cruise-maps-live/data/2023_leg01_OES/gpx/test.gpx')

}
