% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decdeg2degmin.R
\name{decdeg2degmin}
\alias{decdeg2degmin}
\title{decdeg2degmin}
\usage{
decdeg2degmin(decdeg)
}
\arguments{
\item{decdeg}{N-by-1 vector of coordinates in decimal degrees
e.g., decdeg = 30.48667 OR decdeg = c(30.48667, -155.61496)}
}
\value{
N-by-2 data.frame of coordinates with columns 'deg' for
degrees and 'min' for decimal minutes
}
\description{
Utility to convert latitude and longitude coordinates from
decimal degrees to degrees decimal minutes
}
\examples{
# single input coordinates
decdeg = 30.48667
degmin = decdeg2degmin(decdeg)

# multiple input coordinates
decdeg = c(30.48667, -155.61496)
degmin = decdeg2degmin(decdeg)

######################################################################
}
