% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decdeg2degminsec.R
\name{decdeg2degminsec}
\alias{decdeg2degminsec}
\title{decdeg2degminsec}
\usage{
decdeg2degminsec(decdeg)
}
\arguments{
\item{decdeg}{N-by-1 vector of coordinates in decimal degrees
e.g., \code{decdeg <- 30.48667} OR \code{decdeg <- c(30.48667, -155.61496)}}
}
\value{
N-by-3 data.frame of coordinates with columns 'deg' for
degrees, 'min' for minutes, and 'sec' for decimal seconds
}
\description{
Utility to convert latitude and longitude coordinates from
degrees minutes to degrees minutes second
}
\examples{
# single input coordinates
decdeg <- 30.48667
degminsec <- decdeg2degminsec(decdeg)

# multiple input coordinates
decdeg <- c(30.48667, -155.61496)
degminsec <- decdeg2degminsec(decdeg)

}
\author{
Selene Fregosi
}
