% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degmin2degminsec.R
\name{degmin2degminsec}
\alias{degmin2degminsec}
\title{degmin2degminsec}
\usage{
degmin2degminsec(degmin)
}
\arguments{
\item{degmin}{N-by-2 matrix of coordinates with each column representing
degrees and minutes
e.g., matrix(c(30, 29.2020), nrow = 1, ncol = 2)
e.g., matrix(c(30, 29.2020,
-155, 36.8973),
nrow = 2, ncol = 2, byrow = TRUE)}
}
\value{
N-by-3 data.frame of coordinates with columns 'deg for degrees and
'min' for minutes and 'sec' for decimal seconds
}
\description{
Utility to convert latitude and longitude coordinates from
degrees minutes to degrees minutes seconds
}
\examples{
# single input coordinates
degmin = matrix(c(30, 29.2020), nrow = 1, ncol = 2)
degminsec = degmin2degminsec(degmin)

# multiple input coordinates
degmin = matrix(c(30, 29.2020,
                     -155, 36.8973),
                   nrow = 2, ncol = 2, byrow = TRUE)
degminsec = degmin2degminsec(degmin)

}
\author{
Selene Fregosi
}
