% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degminsec2decdeg.R
\name{degminsec2decdeg}
\alias{degminsec2decdeg}
\title{degminsec2decdeg}
\usage{
degminsec2decdeg(degminsec)
}
\arguments{
\item{degminsec}{N-by-3 matrix of coordinates with each column representing
degrees, minutes, and seconds
e.g., matrix(c(30, 29, 12), nrow = 1, ncol = 3)
e.g., matrix(c(30, 29, 12,
-155, 36, 53.838),
nrow = 2, ncol = 3, byrow = TRUE)}
}
\value{
N-by-1 vector of coordinates in decimal degrees
}
\description{
Utility to convert latitude and longitude coordinates from
degrees minutes seconds to decimal degrees
}
\examples{
# single input coordinates
degminsec = matrix(c(30, 29, 12), nrow = 1, ncol = 3)
decdeg = degminsec2decdeg(degminsec)

# multiple input coordinates
degminsec = matrix(c(30, 29, 12,
                     -155, 36, 53.838),
                   nrow = 2, ncol = 3, byrow = TRUE)
decdeg = degminsec2decdeg(degminsec)

}
