% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGoogleDriveFolder.R
\name{downloadGoogleDriveFolder}
\alias{downloadGoogleDriveFolder}
\title{downloadGoogleDriveFolder}
\usage{
downloadGoogleDriveFolder(folder_id, local_path)
}
\arguments{
\item{folder_id}{Google Drive folder ID, parsed as the 'id' column finding
the drive info using googleddrive::drive_get or using
using googledrive::as_id and the folder URL}

\item{local_path}{full path to the folder to save the downloaded files}
}
\description{
recursively download all files and subdirectories in a google
drive folder
}
\author{
Selene Fregosi
}
