% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_concatenatedSlope.R
\name{er_concatenatedSlope}
\alias{er_concatenatedSlope}
\title{er_concatenatedSlope}
\usage{
er_concatenatedSlope(cl, quantLimit = NULL)
}
\arguments{
\item{cl}{list returned by clickSummary() that has a elements spec and snr}

\item{quantLimit}{optional argument to limit the clicks to a particular
quantile range e.g., only clicks in the 25-75th quantiles, 5-95th, etc.
Default is NULL (all clicks), input in format c(.25, .75)}
}
\value{
'cs' the linear regression model output
}
\description{
Model the slope of peak frequencies of clicks sorted by peak
peak frequency as displayed in a concatenated spectrogram using linear
regression. This function models the slope in two ways (1) for all clicks and
(2) for just clicks within the 25-75th percentiles of the mean.
}
\examples{
# all clicks
\dontrun{cs <- concatenatedSlope(cl)}
# 25-75th quantile only
\dontrun{cs <- concatenatedSlope(cl, c(0.25, 0.75))}
}
