% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_filterClicks.R
\name{er_filterClicks}
\alias{er_filterClicks}
\title{er_filterClicks}
\usage{
er_filterClicks(acSt)
}
\arguments{
\item{acSt}{AcousticStudy object generated with processPgDetections()}
}
\value{
an AcousticStudy object containing only filtered detections
}
\description{
filter all click detections to remove clicks that are
unlikely to be odontocetes and are likely noise. Filter settings were tuned
by Jennifer McCullough, see McCullough et al. 2021

Filter settings:
Remove clicks with 10dB bandwidth < 5 kHz
Remove clicks with peak frequency below 5 kHz and above 80 kHz
Remove clicks with durations of less than 2 ms or greater than 1000 ms
For each individual click detector (classifier), remove clicks with peak
frequencies outside the specified limits (remove erroneous
miss-classifications)

last updated: 31 July 2026
}
\examples{
\dontrun{
detsFiltered <- er_filterClicks(dets)
}

}
\author{
Selene Fregosi \email{selene.fregosi@noaa.gov}
}
