% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_plotClickDurationHist.R
\name{er_plotClickDurationHist}
\alias{er_plotClickDurationHist}
\title{er_plotClickDurationHist}
\usage{
er_plotClickDurationHist(cl)
}
\arguments{
\item{cl}{list returned by er_clickSummary() containing elements
goodClicks (data frame with column duration) and nGoodClicks (count)}
}
\value{
creates a plot
}
\description{
Plot histogram of click durations of high SNR clicks only

last updated: 31 July 2026
}
\examples{
\dontrun{
cl <- er_clickSummary(acSt)
er_plotClickDurationHist(cl)
}

}
\seealso{
\code{\link[=er_clickSummary]{er_clickSummary()}}
}
\author{
Selene Fregosi \email{selene.fregosi@noaa.gov}
}
