% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_plotClickSNRHist.R
\name{er_plotClickSNRHist}
\alias{er_plotClickSNRHist}
\title{er_plotClickSNRHist}
\usage{
er_plotClickSNRHist(cl, snrThreshold)
}
\arguments{
\item{cl}{list returned by clickSummary() that has a elements spec and snr}

\item{snrThreshold}{SNR in dB to plot as the threshold for 'good' clicks}
}
\description{
Plot histogram of click SNR for all clicks with a vertical
dashed line at the specified cutoff
}
\examples{
\dontrun{
er_plotClickSNRHist(cl, 15)
}
}
