% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_plotSpectra.R
\name{er_plotSpectra}
\alias{er_plotSpectra}
\title{er_plotSpectra}
\usage{
er_plotSpectra(
  avgSpec,
  cl,
  refSpecs = NULL,
  xLims = c(0, 100),
  snrThreshold = 15
)
}
\arguments{
\item{avgSpec}{list output of PAMpal::calculateAvereageSpectra that includes
variables freq, avgSpec, avgNoise}

\item{cl}{list returned by clickSummary() that has a elements spec and snr
(if SNR is added as an output to calculateAverageSpectra this can be removed)}

\item{refSpecs}{optional argument with list of reference spectra to include
as overlays in the plot}

\item{xLims}{optional argument for frequency limits. Default is 0 to 100.
Define with format c(0, 100)}

\item{snrThreshold}{optional argument for SNR threshold (in dB) for filtering
clicks used in calculation of median spectrum. Default is 15}
}
\description{
Plot spectra of all clicks - mean, median, noise, and any
specified reference spectra
}
\examples{
\dontrun{
er_plotSpectra(avgSpec, cl, refSpecs = refSpecs, xLims = c(0,100), snrThreshold = 15)
}

}
