% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_plotWhistleBeginFreqHist.R
\name{er_plotWhistleBeginFreqHist}
\alias{er_plotWhistleBeginFreqHist}
\title{er_plotWhistleBeginFreqHist}
\usage{
er_plotWhistleBeginFreqHist(wl)
}
\arguments{
\item{wl}{list returned by whistleSummary() that has a elements wh}
}
\description{
Plot histogram begin frequency of whistles
}
\examples{
\dontrun{
er_plotWhistleBeginFreqHist(wl)
}
}
