% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_plotWhistleContours.R
\name{er_plotWhistleContours}
\alias{er_plotWhistleContours}
\title{er_plotWhistleContours}
\usage{
er_plotWhistleContours(dets, eventUID, xMax = 1.5, yMax = 20)
}
\arguments{
\item{dets}{AcousticStudy containing whistle detections}

\item{eventUID}{unique ID for the event being processed (to reference within
the AcousticStudy)}

\item{xMax}{optional argument for upper limit of x axis, in seconds. Default
is 1.5 seconds}

\item{yMax}{optional argument for upper limit of y axis, in kHz. Default is
20 kHz}
}
\description{
Plot all whistle contours
}
\examples{
\dontrun{
er_plotWhistleContours(dets, eventUID, xMax = 1.5, yMax = 20)
}
}
