% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_plotWhistleMedFreqHist.R
\name{er_plotWhistleMedFreqHist}
\alias{er_plotWhistleMedFreqHist}
\title{er_plotWhistleMedFreqHist}
\usage{
er_plotWhistleMedFreqHist(wl)
}
\arguments{
\item{wl}{list returned by whistleSummary() that has a elements wh}
}
\description{
Plot histogram median frequency of whistles
}
\examples{
\dontrun{
er_plotWhistleMedFreqHist(wl)
}
}
