% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDASTrack.R
\name{extractDASTrack}
\alias{extractDASTrack}
\title{extractDASTrack}
\usage{
extractDASTrack(df_proc)
}
\arguments{
\item{df_proc}{processed das file (with swfscDAS::das_process)}
}
\value{
a dataframe of effort tracks with date and lat/lon
}
\description{
Pull effort tracks from a relatively raw daily .das file into
a simpler dataframe. Utilizes the package 'swfscDAS' and then cleans up those
outputs a bit. Generalized from cruise-maps-live's extractTrack()
}
\author{
Selene Fregosi
}
