% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flac2wav.R
\name{flac2wav}
\alias{flac2wav}
\title{flac2wav}
\usage{
flac2wav(path_flac, inDir, outDir)
}
\arguments{
\item{path_flac}{Character. Path to flac executable
e.g., 'C:\flac-1.5.0-win\Win64\flac.exe'}

\item{inDir}{Character. Path to the folder containing FLAC files}

\item{outDir}{Character. Path to the folder where WAV files will be saved.
\emph{This path MUST end in a final slash. A check will ensure that it does.}}
}
\value{
None. Writes WAV files to disk.
}
\description{
Utility to decode FLAC files to WAV format using the command
line flac.exe encoder.

This function converts all .flac files in a directory to .wav using the
flac.exe encoder.
}
\examples{
flac2wav('C:\\\\users\\\\user.name\\\\programs\\\\flac-1.5.0-win\\\\Win64\\\\flac',
'F:\\\\flacFiles', 'F:\\\\wavFiles\\\\');

}
\seealso{
\code{\link[tcltk]{tk_choose.dir}}
Note: This function uses \code{tcltk::tk_choose.dir()} which may require a
working installation of Tcl/Tk.
On Linux and macOS, you may need to install Tcl/Tk separately.

\code{\link[crputils]{wav2flac}}
}
\author{
Selene Fregosi
}
