% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSavedPackages.R
\name{installSavedPackages}
\alias{installSavedPackages}
\title{installSavedPackages}
\usage{
installSavedPackages(fileName)
}
\arguments{
\item{fileName}{filename of saved packages. This can be the full file path
with filename or if no path is specified it will look for the file in your
current working directory. You can use \code{getwd()} to check current
working directory)}
}
\description{
Re-install a set of saved packages (created with
\code{saveInstalledPackages()}) after updating R
}
\author{
Yvonne Barkley and Selene Fregosi
}
