% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lldistkm.R
\name{lldistkm}
\alias{lldistkm}
\title{Calculate distance in km between to lat/lons}
\usage{
lldistkm(latlon1, latlon2)
}
\arguments{
\item{latlon1}{1-by-2 matrix of lat/lon coordinates (in decimal degrees)
e.g., \code{matrix(c(30, 29.2020), nrow = 1, ncol = 2)}\cr}

\item{latlon2}{1-by-2 matrix of lat/lon coordinates (in decimal degrees)
e.g., \code{matrix(c(30, 29.2020), nrow = 1, ncol = 2)}\cr}
}
\value{
list with d1km (distance calculated by haversine formula) and d2km
(distance calculated based on Pythagorean theorem)
}
\description{
Utility to calculate the distance (in kilometers) between a pair
of latitude/longitude coordinates. It returns two distances: d1km which is
calculated using the Haversine formula which accounts for the shape of the
earth and d2km based on Pythagoras' theorem and is not accurate over long
distances. Over short distances d1km and d2km will be very similar.
}
\examples{
#short distance
#create input coordinates
ll1 <- c(21, -157)
ll2 <- c(22, -158)
#calculate distance
dkm <- lldistkm(ll1, ll2)
dkm
# $d1km
# [1] 151.8795
#
# $d2km
# [1] 151.8809

#long distance
#create input coordinates
ll1 <- c(21, -157)
ll2 <- c(13, 144)
#calculate distance
dkm <- lldistkm(ll1, ll2)
dkm
# $d1km
# [1] 6304.003
#
# $d2km
# [1] 32019.57
#d2km not acccurate at longer distances
}
\author{
Selene Fregosi
}
