% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMultiBinaries.R
\name{loadMultiBinaries}
\alias{loadMultiBinaries}
\title{loadMultiBinaries}
\usage{
loadMultiBinaries(wmBinFiles)
}
\arguments{
\item{wmBinFiles}{character array of whistle moan binary files to be loaded}
}
\value{
binOut a list of binary WMD data
}
\description{
Load in multiple Pamguard binary files and compile the data
lists into a single list. For now it works with WMD data...
}
\examples{
\dontrun{
# map the needed binary files
binFiles <- dets@events[[eventUID]]@files$binaries
wmFileIdx <- grep(pattern = '^.*WhistlesMoans_Whistle_and_Moan_Detector_Contours.*\\\\.pgdf$',
                  binFiles)
wmFiles <- dets@events[[eventUID]]@files$binaries[wmFileIdx]
# now load the binaries
loadMultiBinaries(wmFiles)
}
}
