% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASTrackToGPX.R
\name{DASTrackToGPX}
\alias{DASTrackToGPX}
\title{DASTrackToGPX}
\usage{
DASTrackToGPX(et, outGPX)
}
\arguments{
\item{et}{data.frame of effort as tracks, can be 'et' cumulative over a
survey or 'et' for just a single DAS}

\item{outGPX}{fullpath filename to save
example: \code{outGPX <- paste0('newEffortTracks_', dasName, '_', Sys.Date(), '.gpx')}}
}
\value{
none, will write a file
}
\description{
create a .gpx file based on the effort track data recorded in
.DAS files. The track data needs to have been extracted from the
.DAS file using \code{extractDASTrack()} which creates the 'et' dataframe

Track segments within days are plotted separately but are connected. Tracks
across days are not connected.

This was generalized from the cruise-maps-live repository's \code{trackToGPX()}
}
\author{
Selene Fregosi
}
