% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degminsec2degmin.R
\name{degminsec2degmin}
\alias{degminsec2degmin}
\title{degminsec2degmin}
\usage{
degminsec2degmin(degminsec)
}
\arguments{
\item{degminsec}{N-by-3 matrix of coordinates with each column representing
degrees, minutes, and seconds
e.g., matrix(c(30, 29, 12), nrow = 1, ncol = 3)
e.g., matrix(c(30, 29, 12,
-155, 36, 53.838),
nrow = 2, ncol = 3, byrow = TRUE)}
}
\value{
degmin  N-by-2 vector of coordinates in degrees minutes
}
\description{
Utility to convert latitude and longitude coordinates from
degrees minutes seconds to degrees minutes

author: Selene Fregosi selene.fregosi \link{at} noaa.gov
date: 28 July 2023
}
\examples{
# # single input coordinates
degminsec = matrix(c(30, 29, 12), nrow = 1, ncol = 3)
degmin = degminsec2degmin(degminsec)

# # multiple input coordinates
degminsec = matrix(c(30, 29, 12,        
                     -155, 36, 53.838), 
                   nrow = 2, ncol = 3, byrow = TRUE)
degmin = degminsec2degmin(degminsec)

######################################################################
}
