% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_plotClickDurationHist.R
\name{er_plotClickDurationHist}
\alias{er_plotClickDurationHist}
\title{er_plotClickDurationHist}
\usage{
er_plotClickDurationHist(cl)
}
\arguments{
\item{cl}{list returned by clickSummary() that has a elements spec and snr}
}
\description{
Plot histogram of click durations of high SNR clicks only
}
\examples{
\dontrun{
er_plotClickDurationHist
}
}
